backend/products/serializer.py
from rest_framework import serializers


from .models import Product

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = (
            'price',
            'name',
            'category',
            'vehicle_type',
        )

class ProductAdminSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'